Update AutonGUI2.02WORKING
The reason That Steven's GUI (www.github.com/elander1x)did not work is because in his GUI methods, he had a while loop that started with an if statement, making the brain hang until that if statement condition was fufilled. By adding a line of code before that (In this case the allMotorsStop() method), we were able to go to the driver code without having to fufill that if statement.